A S T H E T I C S
Story title is now clickable,  the window now displays without
decoration,  and FINALLY the image (which IS still place holder btw)  is
now displayed when condensed.

Problems atm:  Still deciding image fetching method,  windows should
open in default browser (not whatever its opening in now)  and the
window should be resizable.  Not sure exactly what the issue is
preventing that.

Sidenote:  This is the first version that I consider functioning and
basically useable.   So,  enjoy.